add .propTypes to promo code form
